+++ src/main/java/org/tradingsignal/pojo/yahoo/Split.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Split {
-    private int date;
+    private long date;
     private double numerator;
     private double denominator;
     private String splitRatio;
